+++ asgi_lua/__init__.py
@@ -4,6 +4,7 @@ __version__ = "0.1.0"
 
 from .bundle import pack_bundle, test_bundle, validate_bundle
 from .middleware import LuaConfig, LuaMiddleware
+from .presets import copy_builtin_preset, list_builtin_presets
 from .promotion import compare_decisions, diff_policies
 from .runtime import LuaDecisionError, LuaDecisionTrace, LuaRuntimeError
 from .simulator import simulate_policy
@@ -30,7 +31,9 @@ __all__ = [
     "StateLimits",
     "__version__",
     "compare_decisions",
+    "copy_builtin_preset",
     "diff_policies",
+    "list_builtin_presets",
     "pack_bundle",
     "simulate_policy",
     "test_bundle",